% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAIP_Generator.R
\name{addIdParameter}
\alias{addIdParameter}
\title{addIdParameter
Adds an event-unique Id-parameter to a given flowFrame}
\usage{
addIdParameter(ff)
}
\arguments{
\item{ff}{flowFrame-object}
}
\value{
flowFrame-object with added Id-parameter
}
\description{
addIdParameter
Adds an event-unique Id-parameter to a given flowFrame
}
\examples{
addIdParameter(ff_fcs)
}
