% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAIP_Generator.R
\name{addLaipParameter}
\alias{addLaipParameter}
\title{addLaipParameter
Adds the parameter isLAIP to the given flowFrame.}
\usage{
addLaipParameter(ff)
}
\arguments{
\item{ff}{= flowFrame-object}
}
\value{

}
\description{
addLaipParameter
Adds the parameter isLAIP to the given flowFrame.
}
