% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAIP_Generator.R
\name{log_transform_no_neg}
\alias{log_transform_no_neg}
\title{log_transform_no_neg
Log-transformation that returns zero for all values < 1.}
\usage{
log_transform_no_neg(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
log-transformed numeric vector
}
\description{
log_transform_no_neg
Log-transformation that returns zero for all values < 1.
}
\examples{
log_transform_no_neg(y_axis)
}
