% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAIP_Generator.R
\name{simulate_LAIP_events}
\alias{simulate_LAIP_events}
\title{simulate_LAIP_events}
\usage{
simulate_LAIP_events(
  LAIP,
  blast_pop,
  pop_colnum,
  number_of_LAIPevents,
  LAIP_constraints = NULL,
  addNoise = TRUE,
  printScatterplot = FALSE
)
}
\arguments{
\item{LAIP}{matrix}

\item{blast_pop}{data.frame}

\item{pop_colnum}{number of columns to use in given blast_pop-data.frame}

\item{number_of_LAIPevents}{number of events to be simulated}

\item{LAIP_constraints}{additional constraints for given blasts. blast-events that do not fullfill}

\item{addNoise}{boolean}

\item{printScatterplot}{boolean}
}
\value{
data.frame
}
\description{
Simulates blast-events with given LAIP-expression
}
