% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAIP_Generator.R
\name{simulate_MRD}
\alias{simulate_MRD}
\title{simulate_MRD}
\usage{
simulate_MRD(
  ff,
  LAIP1,
  LAIP2 = NULL,
  LAIP3 = NULL,
  LAIP4 = NULL,
  blast_pop,
  number_of_LAIPevents,
  maxRows = 0,
  LAIP_constraints = NULL,
  addNoise = TRUE,
  noiseIntensity = 0.02,
  addLAIPlabel = TRUE,
  laipLabel = 1e+05,
  printScatterplot = FALSE
)
}
\arguments{
\item{ff}{flowFrame-object to be enriched with LAIP-positive events}

\item{LAIP1}{matrix; specifies parameters to be replaced using expression from given population file}

\item{LAIP2}{matrix (default: NULL); specifies parameters to be replaced using expression from given population file}

\item{LAIP3}{matrix (default: NULL); specifies parameters to be replaced using expression from given population file}

\item{LAIP4}{matrix (default: NULL); specifies parameters to be replaced using expression from given population file}

\item{blast_pop}{data.frame containing blast events}

\item{number_of_LAIPevents}{number of events to be simulated}

\item{maxRows}{numeric (default: 0); specifies number of events in the new MRD-positive flowFrame-object; 0 = unlimited}

\item{LAIP_constraints}{matrix (default: NULL); optional additional constraints for given blasts. Only blast-events that fulfill the constraints are used.}

\item{addNoise}{logical (default: TRUE); modifies blast-data slightly}

\item{noiseIntensity}{numeric (default: 0.02); simulatedValue = (1 +- noiseIntensity) * real value}

\item{addLAIPlabel}{logical (default: TRUE); add isLAIP as additional parameter to given flowFrame labeling the simulated events}

\item{laipLabel}{numeric (default: 100000); specifies the value used to label simulated data. if > 1 LAIP-combination is simulated, the label will be LAIP-number * laipLabel}

\item{printScatterplot}{logical (default: FALSE); print scatterplots of SSC-H and the selected LAIP-parameters}
}
\value{
data.frame containing the simulated LAIP-positive blast events
}
\description{
simulate_MRD is a wrapper function that bundles simulation and integration
of LAIP-positive blasts in a given flowFrame-object.
These events are simulated based on given blast-events with defined
parameters being replaced with expressions of given cell populations.
The simulation of up to 4 different LAIP-populations is supported.
}
\examples{
In this example a bone matter sample with 2 m events is enriched
with two populations of immature blasts with an aberrant expression.
Both LAIP-population will consist of 2000 events each.
The first LAIP-population has an aberrant expression of CD19 and CD56
(LAIP1: CD19+ CD56+), the second LAIP-population shows a loss of HLA_DR
(LAIP2: HLA_DR-).
The data of the bone matter sample are stored in a FCS-file.
The blast events are stored in a csv-file.

library(flowCore)
library(data.table)
fcs_file <- "bm_sample.fcs"
ff_fcs <- read.FCS(fcs_file, transformation = FALSE, truncate_max_range = FALSE)
blast_file <- "blasts.csv"
blast_data <- fread(blast_file)

pro_B_cells are CD19+, nk-cells are CD56+
In this example CD19 is colored with APC-A and CD56 is colored with BV711-A.
To simulate the desired LAIP-population with CD19+ CD56+ the values of APC-A
and BV711-A will be overwritten using values of a range found in pro_B_cells
and nk-cells:

pro_B_cells_file <- "pro_b_cell.csv"
nk_cells_file <- "nk_cell.csv"
LAIP1 <- matrix(c("APC-A",pro_B_cells_file,
                 "BV711-A",nk_cells_file),
               ncol=2,
               byrow=TRUE)

The second LAIP-population (HLA_DR-) is simulated by overwriting the values
of APC-H7-A using values of a range found in granulocytes.

granu_file <- "granulocytes.csv"
LAIP2 <- matrix(c("APC-H7-A",granu_file),
                ncol=2,
                byrow=TRUE)

Only immature blasts shall be used to simulate MRD. This is ensured by setting
corresponding constraints:

LAIP_constraints <- matrix(c("BB700-A",10000,30000, #CD34+
                             "PE-Cy7-A", 5000, NA,  #CD117+,
                             "PEVio615-A",500,2500, #CD133+,
                             "BV421-A",8000,NA,     #CD13+,
                             "APC-R700-A",6000,NA,  #CD33+
                             "APC-H7-A",18000,NA,   #HLADR+
                             "BV786-A",NA,4000,     #CD7-
                             "BV711-A",NA,3000,     #CD56-
                             "FITC-A",NA,3000,      #CD2-
                             "BV650-A",NA,3000,     #CD14-
                             "PerCP-eFluor 710-A",NA,2000,  #CD15-
                             "APC-A",NA,3000,       #CD19-
                             "PE-A",NA,3000,        #CD22-
                             "BV750-A",NA,5000),    #CD11b-
                            ncol=3, byrow=TRUE)
ff_MRD <- simulate_MRD(ff=ff_fcs, LAIP1=LAIP1, LAIP2=LAIP2, blast_pop=blast_data, number_of_LAIPevents=2000, maxRows=2000000, LAIP_constraints=LAIP_constraints)

The new flowFrame-object can be exported into a FCS-file:

output_FileName <- "MRD_positive_bm_sample.fcs"
write.FCS(ff_fcs, output_FileName, what="numeric", delimiter = "|", endian="big")
}
