% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAIP_Generator.R
\name{writeToFCS}
\alias{writeToFCS}
\title{writeToFCS
writeToFCS replaces the data matrix of a given flowFrame with the given data matrix
and saves the flowFrame as FCS-file.}
\usage{
writeToFCS(ff, data, filepath, reshuffle = TRUE)
}
\arguments{
\item{ff}{flowFrame-object to be written into FCS-file}

\item{data}{data.frame or matrix}

\item{filepath}{character(1) string containing the file path of the FCS-file}

\item{reshuffle}{= logical (default: TRUE); if TRUE the rows of data are reshuffled}
}
\description{
writeToFCS
writeToFCS replaces the data matrix of a given flowFrame with the given data matrix
and saves the flowFrame as FCS-file.
}
\examples{
writeToFCS(ff_fcs, data, "MRDpos_BM_sample.fcs", reshuffle = TRUE)
}
