% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAIP_Generator.R
\name{print_scatterplot}
\alias{print_scatterplot}
\title{print_scatterplot
print_scatterplot is a function for debugging purpose}
\usage{
print_scatterplot(
  data,
  max_rows = 10000,
  x = "SSC-H",
  y,
  log_transform_y = TRUE,
  laipColumnName = NULL
)
}
\arguments{
\item{x}{= x-axis}

\item{y}{= y-axis}

\item{laipColumnName}{}
}
\value{

}
\description{
print_scatterplot
print_scatterplot is a function for debugging purpose
}
