% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAIP_Generator.R
\name{writeToFCS}
\alias{writeToFCS}
\title{writeToFCS
writeToFCS replaces the data matrix of a given flowFrame with the given data matrix
and saves the flowFrame as FCS-file.}
\usage{
writeToFCS(ff, data, filepath, reshuffle = TRUE)
}
\arguments{
\item{ff}{= flowFrame-object}

\item{data}{= matrix}

\item{reshuffle}{= boolean, if TRUE the rows of data are reshuffled}
}
\value{

}
\description{
writeToFCS
writeToFCS replaces the data matrix of a given flowFrame with the given data matrix
and saves the flowFrame as FCS-file.
}
